+++ packages/google_atlas/example/main.dart
@@ -18,7 +18,7 @@ class MyApp extends StatelessWidget {
 }
 
 class AtlasSample extends StatelessWidget {
-  final CameraPosition _initialCameraPosition = CameraPosition(
+  final CameraPosition cameraPosition = CameraPosition(
     target: LatLng(
       latitude: 37.42796133580664,
       longitude: -122.085749655962,
@@ -29,7 +29,7 @@ class AtlasSample extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       body: Atlas(
-        initialCameraPosition: _initialCameraPosition,
+        cameraPosition: cameraPosition,
       ),
     );
   }